refactor: migrate platformGrid to CSS modules
