+++ src/components/platformGrid/styles.module.scss
@@ -1,0 +1,60 @@
+.PlatformCell {
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+
+  overflow: hidden;
+
+  background: #fff;
+  color: purple;
+  border: 1px solid #fff;
+  border-radius: 0.5rem;
+  padding: 1rem 1.5rem;
+  line-height: 1.5;
+  transition-property: transform, box-shadow, border-color;
+  transition-duration: 0.25s;
+  transition-timing-function: ease-out;
+  box-shadow: 0px 2px 4px rgba(64, 30, 76, 0.1);
+}
+
+.PlatformCellIcon {
+  width: 82px;
+  height: 82px;
+}
+
+.PlatformCellContent {
+  flex: 1;
+  margin-left: 1.5rem;
+
+  h4 {
+    margin: 0 0 0.5rem;
+  }
+}
+
+.GuideList {
+  font-size: 0.8em;
+  width: 100%;
+  text-overflow: ellipsis;
+
+  a {
+    display: inline-block;
+
+    &:after {
+      content: ', ';
+      padding: 0 5px 0 0;
+    }
+
+    &:last-child:after {
+      content: '';
+      margin: 0;
+    }
+  }
+}
+
+.PlatformGrid {
+  display: grid;
+  gap: 1.5rem;
+  grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+  margin-bottom: 1.5rem;
+  max-width: 56rem;
+}
